Release version 7.15.0

+++ Applozic/DomainClasses/ALContact.m
@@ -120,7 +120,7 @@
 }
 
 -(BOOL)isDeleted {
-    return self.deletedAtTime != nil && self.deletedAtTime.intValue > 0;
+    return self.deletedAtTime != nil && self.deletedAtTime.longValue > 0;
 }
 
 @end
